# contracts/openapi/ai-control-plane-v1.yaml
openapi: 3.1.0
info:
  title: AI ERP Control Plane API
  version: 1.5.0
  description: >-
    Site-scoped, draft-only AI proposal boundary. The control plane never has
    credentials for an ERP database and never receives permission to post ERP
    transactions.
paths:
  /healthz:
    get:
      summary: Report control-plane process health
      security: []
      responses:
        '200':
          description: The control-plane process is healthy.
  /readyz:
    get:
      summary: Report approved provider configuration readiness
      security: []
      responses:
        '200':
          description: The configured provider is approved and ready for requests.
        '503':
          description: The configured provider is missing or not approved.
  /v1/proposals/service-closeout-summary:
    post:
      summary: Draft a cited service-work-order closeout summary
      operationId: draftServiceCloseoutSummary
      security:
        - ControlPlaneServiceKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCloseoutSummaryRequest'
      responses:
        '200':
          description: A draft-only proposal. It has no ERP side effect.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalResponse'
        '401':
          description: Missing or invalid service credential.
        '422':
          description: Invalid or unsupported request payload.
        '503':
          description: The selected approved model provider is unavailable.
  /v1/proposals/scheduling-explanation:
    post:
      summary: Draft a deterministic scheduling-suggestion explanation
      operationId: draftSchedulingExplanation
      security:
        - ControlPlaneServiceKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulingExplanationRequest'
      responses:
        '200':
          description: A draft-only explanation. It has no ERP side effect.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulingProposalResponse'
        '401':
          description: Missing or invalid service credential.
        '422':
          description: Invalid or unsupported request payload.
  /v1/proposals/exception-recovery:
    post:
      summary: Draft deterministic recovery steps for an open closure exception
      operationId: draftExceptionRecovery
      security:
        - ControlPlaneServiceKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExceptionRecoveryRequest'
      responses:
        '200':
          description: A draft-only recovery proposal. It has no ERP side effect.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionRecoveryProposalResponse'
        '401':
          description: Missing or invalid service credential.
        '422':
          description: Invalid or unsupported request payload.
  /v1/proposals/repair-memory:
    post:
      summary: Draft cited repair memory from prior completed work
      operationId: draftRepairMemory
      security:
        - ControlPlaneServiceKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepairMemoryRequest'
      responses:
        '200':
          description: A draft-only repair-memory proposal. It has no ERP side effect.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairMemoryProposalResponse'
        '401':
          description: Missing or invalid service credential.
        '422':
          description: Invalid or unsupported request payload.
components:
  securitySchemes:
    ControlPlaneServiceKey:
      type: http
      scheme: bearer
      bearerFormat: opaque-service-key
  schemas:
    SourceReference:
      type: object
      additionalProperties: false
      required: [doctype, name, field, content_hash]
      properties:
        doctype:
          type: string
          minLength: 1
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 256
        field:
          type: string
          minLength: 1
          maxLength: 128
        content_hash:
          type: string
          pattern: '^[a-f0-9]{64}$'
    TimeEntry:
      type: object
      additionalProperties: false
      required: [technician, work_date, time_type, hours]
      properties:
        technician:
          type: string
          minLength: 1
          maxLength: 256
        work_date:
          type: string
          format: date
        time_type:
          type: string
          minLength: 1
          maxLength: 128
        hours:
          type: number
          exclusiveMinimum: 0
          maximum: 24
    RelatedWorkSummary:
      type: object
      additionalProperties: false
      required: [name, subject, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 256
        subject:
          type: string
          minLength: 1
          maxLength: 256
        status:
          type: string
          minLength: 1
          maxLength: 128
        inspection_result:
          type: string
          maxLength: 128
        closeout_notes:
          type: string
          maxLength: 4000
    PartUsage:
      type: object
      additionalProperties: false
      required: [item, qty, source_warehouse, issued]
      properties:
        item:
          type: string
          minLength: 1
          maxLength: 256
        qty:
          type: number
          exclusiveMinimum: 0
          maximum: 100000
        source_warehouse:
          type: string
          minLength: 1
          maxLength: 256
        issued:
          type: boolean
    ServiceCloseoutSummaryRequest:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - request_id
        - tenant_site
        - requested_by
        - work_order
        - sources
      properties:
        schema_version:
          type: integer
          const: 1
        request_id:
          type: string
          format: uuid
        tenant_site:
          type: string
          minLength: 1
          maxLength: 253
        requested_by:
          type: string
          minLength: 1
          maxLength: 256
        work_order:
          type: object
          additionalProperties: false
          required: [doctype, name, subject, status, time_entries, parts]
          properties:
            doctype:
              type: string
              const: Service Work Order
            name:
              type: string
              minLength: 1
              maxLength: 256
            subject:
              type: string
              minLength: 1
              maxLength: 256
            status:
              type: string
              minLength: 1
              maxLength: 128
            description:
              type: string
              maxLength: 4000
            closeout_notes:
              type: string
              maxLength: 4000
            time_entries:
              type: array
              maxItems: 100
              items:
                $ref: '#/components/schemas/TimeEntry'
            parts:
              type: array
              maxItems: 200
              items:
                $ref: '#/components/schemas/PartUsage'
            related_history:
              type: array
              maxItems: 5
              items:
                $ref: '#/components/schemas/RelatedWorkSummary'
        sources:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/SourceReference'
    ProposalResponse:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - request_id
        - proposal_type
        - policy
        - model
        - draft_content
        - sources
      properties:
        schema_version:
          type: integer
          const: 1
        request_id:
          type: string
          format: uuid
        proposal_type:
          type: string
          const: service_closeout_summary
        policy:
          type: object
          additionalProperties: false
          required: [decision, allowed_action, reason]
          properties:
            decision:
              type: string
              const: draft_only
            allowed_action:
              type: string
              const: none
            reason:
              type: string
        model:
          type: object
          additionalProperties: false
          required: [provider, name, prompt_version]
          properties:
            provider:
              type: string
            name:
              type: string
            prompt_version:
              type: string
        audit:
          anyOf:
            - $ref: '#/components/schemas/ProviderAudit'
            - type: 'null'
        draft_content:
          type: string
          minLength: 1
          maxLength: 8000
        sources:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SourceReference'
    SchedulingCandidate:
      type: object
      additionalProperties: false
      required: [technician, score, workload, familiarity, reasons]
      properties:
        technician:
          type: string
          minLength: 1
          maxLength: 256
        score:
          type: integer
        workload:
          type: integer
          minimum: 0
        familiarity:
          type: integer
          minimum: 0
        reasons:
          type: array
          maxItems: 10
          items:
            type: string
            maxLength: 64
    SchedulingExclusion:
      type: object
      additionalProperties: false
      required: [technician, reason]
      properties:
        technician:
          type: string
          minLength: 1
          maxLength: 256
        reason:
          type: string
          minLength: 1
          maxLength: 128
    SchedulingExplanationRequest:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - request_id
        - tenant_site
        - requested_by
        - work_order
        - candidates
        - sources
      properties:
        schema_version:
          type: integer
          const: 1
        request_id:
          type: string
          format: uuid
        tenant_site:
          type: string
          minLength: 1
          maxLength: 253
        requested_by:
          type: string
          minLength: 1
          maxLength: 256
        work_order:
          type: object
          additionalProperties: false
          required: [doctype, name, subject, status]
          properties:
            doctype:
              type: string
              const: Service Work Order
            name:
              type: string
              minLength: 1
              maxLength: 256
            subject:
              type: string
              minLength: 1
              maxLength: 256
            status:
              type: string
              minLength: 1
              maxLength: 128
            service_priority:
              type: string
              maxLength: 128
            sla_due_at:
              type: string
              maxLength: 64
        candidates:
          type: array
          maxItems: 5
          items:
            $ref: '#/components/schemas/SchedulingCandidate'
        excluded:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/SchedulingExclusion'
        sources:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/SourceReference'
    SchedulingProposalResponse:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - request_id
        - proposal_type
        - policy
        - model
        - draft_content
        - sources
      properties:
        schema_version:
          type: integer
          const: 1
        request_id:
          type: string
          format: uuid
        proposal_type:
          type: string
          const: scheduling_explanation
        policy:
          type: object
          additionalProperties: false
          required: [decision, allowed_action, reason]
          properties:
            decision:
              type: string
              const: draft_only
            allowed_action:
              type: string
              const: none
            reason:
              type: string
        model:
          type: object
          additionalProperties: false
          required: [provider, name, prompt_version]
          properties:
            provider:
              type: string
            name:
              type: string
            prompt_version:
              type: string
        draft_content:
          type: string
          minLength: 1
          maxLength: 8000
        sources:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SourceReference'
    RecoveryPart:
      type: object
      additionalProperties: false
      required: [item, qty, issued]
      properties:
        item:
          type: string
          minLength: 1
          maxLength: 256
        qty:
          type: number
          exclusiveMinimum: 0
          maximum: 100000
        issued:
          type: boolean
    ClosureExceptionSummary:
      type: object
      additionalProperties: false
      required: [name, reason, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 256
        reason:
          type: string
          minLength: 1
          maxLength: 128
        status:
          type: string
          minLength: 1
          maxLength: 128
        due_date:
          type: string
          maxLength: 64
    ExceptionRecoveryRequest:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - request_id
        - tenant_site
        - requested_by
        - work_order
        - exception
        - sources
      properties:
        schema_version:
          type: integer
          const: 1
        request_id:
          type: string
          format: uuid
        tenant_site:
          type: string
          minLength: 1
          maxLength: 253
        requested_by:
          type: string
          minLength: 1
          maxLength: 256
        work_order:
          type: object
          additionalProperties: false
          required: [doctype, name, subject, status]
          properties:
            doctype:
              type: string
              const: Service Work Order
            name:
              type: string
              minLength: 1
              maxLength: 256
            subject:
              type: string
              minLength: 1
              maxLength: 256
            status:
              type: string
              minLength: 1
              maxLength: 128
            cannot_close_reason:
              type: string
              maxLength: 128
            inspection_result:
              type: string
              maxLength: 128
        exception:
          $ref: '#/components/schemas/ClosureExceptionSummary'
        parts:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/RecoveryPart'
        related_history:
          type: array
          maxItems: 5
          items:
            $ref: '#/components/schemas/RelatedWorkSummary'
        sources:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/SourceReference'
    ExceptionRecoveryProposalResponse:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - request_id
        - proposal_type
        - policy
        - model
        - draft_content
        - sources
      properties:
        schema_version:
          type: integer
          const: 1
        request_id:
          type: string
          format: uuid
        proposal_type:
          type: string
          const: exception_recovery
        policy:
          type: object
          additionalProperties: false
          required: [decision, allowed_action, reason]
          properties:
            decision:
              type: string
              const: draft_only
            allowed_action:
              type: string
              const: none
            reason:
              type: string
        model:
          type: object
          additionalProperties: false
          required: [provider, name, prompt_version]
          properties:
            provider:
              type: string
            name:
              type: string
            prompt_version:
              type: string
        draft_content:
          type: string
          minLength: 1
          maxLength: 8000
        sources:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SourceReference'
    RepairHistoryEntry:
      type: object
      additionalProperties: false
      required: [name, subject, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 256
        subject:
          type: string
          minLength: 1
          maxLength: 256
        status:
          type: string
          minLength: 1
          maxLength: 128
        inspection_result:
          type: string
          maxLength: 128
        closeout_notes:
          type: string
          maxLength: 4000
        parts:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/RecoveryPart'
    RepairMemoryRequest:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - request_id
        - tenant_site
        - requested_by
        - work_order
        - sources
      properties:
        schema_version:
          type: integer
          const: 1
        request_id:
          type: string
          format: uuid
        tenant_site:
          type: string
          minLength: 1
          maxLength: 253
        requested_by:
          type: string
          minLength: 1
          maxLength: 256
        work_order:
          type: object
          additionalProperties: false
          required: [doctype, name, subject, status]
          properties:
            doctype:
              type: string
              const: Service Work Order
            name:
              type: string
              minLength: 1
              maxLength: 256
            subject:
              type: string
              minLength: 1
              maxLength: 256
            status:
              type: string
              minLength: 1
              maxLength: 128
            description:
              type: string
              maxLength: 4000
        related_history:
          type: array
          maxItems: 5
          items:
            $ref: '#/components/schemas/RepairHistoryEntry'
        sources:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/SourceReference'
    RepairMemoryProposalResponse:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - request_id
        - proposal_type
        - policy
        - model
        - draft_content
        - sources
      properties:
        schema_version:
          type: integer
          const: 1
        request_id:
          type: string
          format: uuid
        proposal_type:
          type: string
          const: repair_memory
        policy:
          type: object
          additionalProperties: false
          required: [decision, allowed_action, reason]
          properties:
            decision:
              type: string
              const: draft_only
            allowed_action:
              type: string
              const: none
            reason:
              type: string
        model:
          type: object
          additionalProperties: false
          required: [provider, name, prompt_version]
          properties:
            provider:
              type: string
            name:
              type: string
            prompt_version:
              type: string
        audit:
          anyOf:
            - $ref: '#/components/schemas/ProviderAudit'
            - type: 'null'
        draft_content:
          type: string
          minLength: 1
          maxLength: 8000
        sources:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SourceReference'
    ProviderAudit:
      type: object
      additionalProperties: false
      required: [response_id_hash, input_tokens, output_tokens, duration_ms, redaction_count]
      properties:
        response_id_hash:
          type: string
          pattern: '^[a-f0-9]{64}$'
        input_tokens:
          type: integer
          minimum: 0
        output_tokens:
          type: integer
          minimum: 0
        duration_ms:
          type: integer
          minimum: 0
        redaction_count:
          type: integer
          minimum: 0